Improve search routing
* added router and param function to root of search
