+++ pages/search/index.js
@@ -1,47 +1,17 @@
-const { merge } = require('lodash');
-const { page } = require('@asl/service/ui');
-const datatable = require('@asl/pages/pages/common/routers/datatable');
-const schemas = require('./schema');
-const NotFoundError = require('../../lib/errors/not-found');
-const content = require('./content');
+const { Router } = require('express');
+const search = require('./search');
 
-const searchableModels = ['establishments', 'profiles', 'projects'];
+module.exports = () => {
 
-module.exports = settings => {
-  const app = page({
-    ...settings,
-    root: __dirname
-  });
+  const app = Router();
 
-  app.use('/', (req, res, next) => {
-    req.searchType = req.params.searchType || searchableModels[0];
-
-    merge(res.locals.static.content, content(req.searchType));
-    res.locals.static.profile = req.user.profile;
-    res.locals.static.searchType = req.searchType;
-    res.locals.static.searchableModels = req.query.searchableModels;
+  app.param('searchType', (req, res, next, searchType) => {
+    req.searchType = searchType;
     next();
   });
 
-  app.use('/', datatable({
-    configure: (req, res, next) => {
-      const searchType = req.searchType;
-
-      if (searchType && !searchableModels.includes(searchType)) {
-        return next(new NotFoundError());
-      }
-
-      req.datatable.searchType = searchType;
-      req.datatable.schema = schemas[searchType];
-      req.datatable.apiPath = `/search/${searchType}`;
-
-      next();
-    }
-  })());
-
-  app.use((req, res, next) => {
-    res.sendResponse();
-  });
+  app.use('/:searchType', search());
+  app.use('/', search());
 
   return app;
 };
